General CMake and testing and discovery fixes (#375)
* Only set boost target options the target exists

* Fix Boost and GTest alias target issues

* Fix compiling with Boost 1.85 (add Align to BOOST_INCLUDE_LIBRARIES)

* Fix Win32 Visual Studio 2017 compilation

* Disable Warnings As Errors for external dependencies fix test warnings

* Make ModuleManager ignore invalid paths and just print a warning

* Make ModuleManager tests use relative paths instead of fixed/absolute

* Fix CTest absolute working directory paths to relative

* Set working directory to test runner location for `test_module_manager_internals`

* Workaround mDNS discovery no longer providing device IP

* Fix finding only IPv6 addresses or multiple ConfigurationAndStreaming capabilities

* Fix NativeStreamingDevice protocol adding IPv4 address even if there is none

* Add a default argument to changed Context() factory to maintain backwards compatibility

* Add fs::proximate() shim for std::experimental::filesystem

* Refactor presets so they can be included in vendor ones

* Add address type and reachability status

---------

+++ shared/libraries/opcua/opcuashared/tests/test_opcuaendpoint.cpp
@@ -44,7 +44,7 @@ TEST_F(OpcUaEndpointTest, RegisterCustomTypes)
     endpoint.registerCustomTypes(UA_TYPES_COUNT, UA_TYPES);
     auto typeList = endpoint.getCustomDataTypes();
 
-    ASSERT_EQ(typeList[0].typesSize, UA_TYPES_COUNT);
+    ASSERT_EQ(typeList[0].typesSize, static_cast<std::size_t>(UA_TYPES_COUNT));
     ASSERT_EQ(typeList[0].types, UA_TYPES);
 }
